Add user editing and deletion functionality with corresponding UI

+++ admin/adminUserCrud.php
@@ -6,25 +6,19 @@ ini_set('display_errors', '1');
 ini_set('display_startup_errors', '1');
 error_reporting(E_ALL);
 
-
-print '<pre>';
-print_r($_POST);
-print '</pre>';
-
-print '<pre>';
-print_r($errors);
-print '</pre>';
+// print '<pre>';
+// print_r($errors);
+// print '</pre>';
 
 $users = getUser();
 
-if (isset($_GET['id']) && is_numeric($_GET['id'])) {
-    $id = intval($_GET['id']); // Получить ID из URL
+if (isset($_GET['action']) && $_GET['action'] === 'delete' && isset($_GET['id']) && is_numeric($_GET['id'])) {
+    $id = intval($_GET['id']);
     if (deleteUser($id)) {
-        header("Location: index.php?message=Record deleted successfully");
+        header("Location: adminUserCrud.php?message=Record deleted successfully");
         exit;
     } else {
-        echo "Error: Could not delete record.";
-        exit;
+        $errors[] = "Error deleting user.";
     }
 }
 
@@ -36,18 +30,13 @@ if (isset($_GET['id']) && is_numeric($_GET['id'])) {
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>CRUD Products</title>
-    <link rel="stylesheet" href="./css/adminProduct.css">
+    <title>User Administration Panel</title>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css">
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto|Varela+Round">
-    <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.css">
+    <link rel="stylesheet" href="./css/adminProduct.css">
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.4/jquery.min.js"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.css" integrity="sha384-tViUnnbYAV00FLIhhi3v/dWt3Jxw4gZQcNoSCxCIFNJVCx7/D55/wXsrNIRANwdD" crossorigin="anonymous">
-
-
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"></script>
 </head>
 
 <body>
@@ -59,63 +48,89 @@ if (isset($_GET['id']) && is_numeric($_GET['id'])) {
         <section>
             <h2>Users</h2>
 
-            <form method="post" action="./adminUserCrud.php">
+            <?php if (isset($_GET['message'])): ?>
+                <div class="alert alert-success" style="font-size: 1.2rem;">
+                    <?= htmlspecialchars($_GET['message']); ?>
+                </div>
+            <?php endif; ?>
+
+            <?php if (!empty($errors)): ?>
+                <div class=" alert alert-danger" style="font-size: 1.2rem;">
+                    <?php foreach ($errors as $error): ?>
+                        <p><?= htmlspecialchars($error); ?></p>
+                    <?php endforeach; ?>
+                </div>
+            <?php endif; ?>
+
+            <form method="get" action="./adminUserCrud.php">
             </form>
         </section>
-        <a href="addform.php"><button class="btn btn-success btn-lg float-right" type="submit"><i class="bi bi-plus-circle"></i> Add new item</button></a>
+
         <div class="container">
+            <div class="table-wrapper" style="position: relative; margin-top: 60px;">
+                <a href="addform.php">
+                    <button class="btn btn-success btn-lg" type="submit" style="position: absolute; top: -50px; right: 0;">
+                        <i class="bi bi-plus-circle"></i> Add new item
+                    </button>
+                </a>
 
-            <div class="table-responsive">
-                <div class="table-wrapper">
 
-                    <table class="table table-striped table-hover">
-                        <thead>
-                            <tr>
-                                <th scope="col">Id</th>
-                                <th scope="col">First name</th>
-                                <th scope="col">Last name</th>
-                                <th scope="col">Username</th>
-                                <th scope="col">Country</th>
-                                <th scope="col">Email</th>
-                                <th scope="col">Password</th>
-                                <th scope="col">Pet name</th>
-                                <th scope="col">Actions</th>
-                            </tr>
-                        </thead>
-                        <tbody>
+                <div class="table-responsive">
+                    <div class="table-wrapper">
 
-                            <?php if (isset($users) && is_array($users)): ?>
-                                <?php foreach ($users as $user): ?>
-
-                                    <td><?= $user['id'] ?></td>
-                                    <td><?= $user['firstname'] ?></td>
-                                    <td><?= $user['lastname'] ?></td>
-                                    <td><?= $user['username'] ?></td>
-                                    <td><?= $user['country'] ?></td>
-                                    <td><?= $user['mail'] ?></td>
-                                    <td><?= $user['password'] ?></td>
-                                    <td><?= $user['petname'] ?></td>
-                                    <td>
-                                        <a href="edit.php?id=<?= $user['id']; ?>">
-                                            <button type="button" class="btn btn-outline-warning">Edit</button></a>
-                                        <a href="adminUserCrud.php">
-                                            <a href="adminUserCrud.php?action=delete&id=<?php echo $user['id']; ?>"></a> <button type="button" class="btn btn-outline-danger">Delete</button></a>
-                                    </td>
-
-                                    </tr>
-
-                                <?php endforeach; ?>
-                            <?php else: ?>
+                        <table class="table table-striped table-hover">
+                            <thead>
                                 <tr>
-                                    <td colspan="9">No users found.</td>
+                                    <th scope="col">Id</th>
+                                    <th scope="col">First name</th>
+                                    <th scope="col">Last name</th>
+                                    <th scope="col">Username</th>
+                                    <th scope="col">Country</th>
+                                    <th scope="col">Email</th>
+                                    <th scope="col">Password</th>
+                                    <th scope="col">Pet name</th>
+                                    <th scope="col">Actions</th>
                                 </tr>
-                            <?php endif; ?>
+                            </thead>
+                            <tbody>
 
-                        </tbody>
-                    </table>
+                                <?php if (isset($users) && is_array($users)): ?>
+                                    <?php foreach ($users as $user): ?>
+
+                                        <td><?= $user['id'] ?></td>
+                                        <td><?= $user['firstname'] ?></td>
+                                        <td><?= $user['lastname'] ?></td>
+                                        <td><?= $user['username'] ?></td>
+                                        <td><?= $user['country'] ?></td>
+                                        <td><?= $user['mail'] ?></td>
+                                        <td><?= $user['password'] ?></td>
+                                        <td><?= $user['petname'] ?></td>
+                                        <td>
+                                            <a href="editUser.php?id=<?= $user['id']; ?>" class="d-inline-block me-2">
+                                                <button type="button" class="btn btn-outline-warning">Edit</button>
+                                            </a>
+                                        </td>
+                                        <td>
+                                            <a href="deleteUser.php?id=<?= $user['id']; ?>" class="d-inline-block">
+                                                <button type="button" class="btn btn-outline-danger">Delete</button>
+                                            </a>
+                                        </td>
+
+
+                                        </tr>
+
+                                    <?php endforeach; ?>
+                                <?php else: ?>
+                                    <tr>
+                                        <td colspan="9">No users found.</td>
+                                    </tr>
+                                <?php endif; ?>
+
+                            </tbody>
+                        </table>
+                    </div>
                 </div>
             </div>
-        </div>
     </main>
 </body>
